SKU四位小数修改

+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/ProductSkuStockInfoDraft.java
@@ -43,6 +43,6 @@ public class ProductSkuStockInfoDraft extends CommonBean {
     private Integer baseProductContent;
 
     @ApiModelProperty("拆零系数")
-    private BigDecimal zeroRemovalCoefficient;
+    private Long zeroRemovalCoefficient;
 
 }
